chore: upgrade @atpradical/picopico-ui-kit to v0.18.6
Upgraded the `@atpradical/picopico-ui-kit` dependency in `package.json` and updated the `pnpm-lock.yaml` file accordingly. This ensures compatibility with the latest features and bug fixes provided in the `0.18.6` release.

---

### fix: rename prop `items` to `posts` in `Publications` component

Renamed the prop `items` to `posts` in the `Publications` component for better clarity and consistency with data naming conventions. This change requires updating all instances where the `items` prop is used.

---

### feat: add PostDialog component for displaying post details

Introduced the `PostDialog` component to display detailed views of selected posts, including post images and descriptions. Integrated the new dialog within the `Publications` component to enhance user interaction with individual posts.